requirements.txt
fastapi>=0.95.1
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.21
jinja2>=3.1.2
python-dotenv>=0.21.0
python-multipart>=0.0.6
reportlab>=3.6.0
openpyxl>=3.1.2
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
starlette>=0.27.0
pydantic>=1.10.8
bcrypt>=4.0.1
email-validator
num2words
pandas>=1.3.0
openpyxl>=3.0.9
itsdangerous
